Prefixed the category methods to reduce the risk of name collisions.
Categories on classes like NSString in your own codebase are one thing.
Adding them in a bundle that is to be loaded by a behemoth like Xcode, is another.
And since there are no namespaces in ObjC, and collisions in categories are essentially non-deterministic, extra care needs to be taken when doing so.

Hence:
Prefix All the Things!