Add an option for suppressing the blank lines between sections.
The setting is stored in the negative (“no blank line between sections”) so that the default state for existing users is to be on, thus no behavior change. When switched on, VVBaseCommenter’s emptyLine is just an empty string.

Also updated VVDocumenterSetting to use @property instead of manually declaring getters and setters.